Move the files (notation & test) & add Beat
