+++ MinGW/build.cmd
@@ -1,2 +1,2 @@
 @echo off
-g++ -std=c++14 ../Source/main.cpp ../Source/midiplayer.cpp ../Source/parse.cpp -lWinmm -o midiplayer
+g++ -std=c++14 ../Source/main.cpp ../Source/midiplayer.cpp ../Source/notation.cpp ../Source/test.cpp ../Source/parse.cpp -lWinmm -o midiplayer
