+++ Source/parse.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "midiplayer.h"
+#include "notation.h"
 
 // Example : #C4 C#4
 MIDIPlayer::Pitch parsePitch(const char *word, const char *&end);
