--- VS2017/MIDIPlayer/NumberedNotation.txt
@@ -1,17 +1,0 @@
-1 = C
- 1 1#  2 2#  3 3#  4 4#  5 5#  6 6#  7 7# +1
- C #C  D #D  E  F  F #F  G #G  A #A  B  C  C
-3c 3d 3e 3f 40 41 41 42 43 44 45 46 47 48 48
- 0  1  2  3  4  5  5  6  7  8  9 10 11 12 12
-
-1 = #C
- 1 1#  2 2#  3 3#  4 4#  5 5#  6 6#  7 7# +1
-#C  D #D  E  F #F #F  G #G  A #A  B  C #C #C
-3d 3e 3f 40 41 42 42 43 44 45 46 47 48 49 49
- 0  1  2  3  4  5  5  6  7  8  9 10 11 12 12
-
-1 = D
- 1 1#  2 2#  3 3#  4 4#  5 5#  6 6#  7 7# +1
- D #D  E #E #F  G  G #G  A #A  B  C #C  D  D
-3e 3f 40 41 42 43 43 44 45 46 47 48 49 50 50
- 0  1  2  3  4  5  5  6  7  8  9 10 11 12 12
